+++ CHANGELOG.md
@@ -42,6 +42,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     `Required` or `Write`.
   - `Get-CompositeResourceParameterValidateSet` - Returns the array of values
     contained in the ValidateSet parameter attributes if it exists.
+  - `Show-InvokeGitReturn` - Displays `Invoke-Git` returned hashtable
+    via Write-Verbose and Write-Debug. Providing better feedback.
 - Added QA test to do some quality checks on the module code and change log.
 
 ### Changed